Move highlighting and matching_brace
